+++ Business/Commands/GetAllPartsCommand.cs
@@ -1,9 +1,9 @@
 using Data.Dtos;
 using Data.Entities;
-using Data.Infrastructure;
 using Microsoft.EntityFrameworkCore;
 using WF.Cqrs.Handlers;
 using WF.Cqrs.Operations;
+using WF.Data.Relational.Context;
 
 namespace Business.Commands;
