+++ Business/Extensions/ServiceCollectionExtensions.cs
@@ -1,7 +1,7 @@
-using Data.Extensions;
-using Data.Infrastructure;
 using Microsoft.Extensions.DependencyInjection;
 using WF.Cqrs.Extensions;
+using WF.Data.Relational.Context;
+using WF.Data.Relational.Extensions;
 using WF.Mapper;
 using WF.Mapper.Extensions;
 
@@ -10,7 +10,7 @@ namespace Business.Extensions;
 public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddBusiness<TConfigureDbContext>(this IServiceCollection services)
-        where TConfigureDbContext : class, IConfigureDbContext
+        where TConfigureDbContext : class, IDbContextConfigurator
         => services
             .AddData<TConfigureDbContext>()
             .AddMapper<DefaultMapper>(typeof(ServiceCollectionExtensions))
