--- Data/Configurators/PartConfigurator.cs
@@ -1,6 +1,0 @@
-using Data.Configurators.Abstractions;
-using Data.Entities;
-
-namespace Data.Configurators;
-
-public class PartConfigurator : BaseConfigurator<Part>;
