+++ Data/Configurators/PartEntityConfigurator.cs
@@ -1,0 +1,6 @@
+using Data.Entities;
+using WF.Data.Relational.Configurators;
+
+namespace Data.Configurators;
+
+public class PartEntityConfigurator : BaseEntityConfigurator<Part>;
