+++ Data/Data.csproj
@@ -7,17 +7,8 @@
     </PropertyGroup>
 
     <ItemGroup>
-        <PackageReference Include="Microsoft.EntityFrameworkCore.Relational" Version="8.0.1"/>
-        <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="8.0.0"/>
-    </ItemGroup>
-
-    <ItemGroup>
-        <Compile Update="Entities\Abstractions\BaseEntity.cs">
-            <DependentUpon>IEntity.cs</DependentUpon>
-        </Compile>
-        <Compile Update="Dtos\Abstractions\BaseDto.cs">
-            <DependentUpon>IDto.cs</DependentUpon>
-        </Compile>
+        <ProjectReference Include="..\WF.Data.Relational\WF.Data.Relational.csproj"/>
+        <ProjectReference Include="..\WF.Domain\WF.Domain.csproj"/>
     </ItemGroup>
 
 </Project>
