--- Data/Dtos/Abstractions/BaseDto.cs
@@ -1,6 +1,0 @@
-namespace Data.Dtos.Abstractions;
-
-public record BaseDto : IDto
-{
-    public int Id { get; set; }
-}
