--- Data/Dtos/Abstractions/IDto.cs
@@ -1,6 +1,0 @@
-namespace Data.Dtos.Abstractions;
-
-public interface IDto
-{
-    public int Id { get; set; }
-}
