+++ Data/Dtos/PartDto.cs
@@ -1,4 +1,4 @@
-﻿using Data.Dtos.Abstractions;
+﻿using WF.Domain.Dtos;
 
 namespace Data.Dtos;
 
@@ -7,6 +7,7 @@ public record PartDtoSummary : BaseDto
     public required string Type { get; set; } //TODO: make enum
     public required string Location { get; set; }
     public required int Quantity { get; set; }
+
     // ReSharper disable once PropertyCanBeMadeInitOnly.Global
     public required decimal Size1 { get; set; }
     public decimal? Size2 { get; set; }
