--- Data/Entities/Abstractions/BaseEntity.cs
@@ -1,6 +1,0 @@
-namespace Data.Entities.Abstractions;
-
-public record BaseEntity : IEntity
-{
-    public int Id { get; set; }
-}
