--- Data/Entities/Abstractions/IEntity.cs
@@ -1,6 +1,0 @@
-namespace Data.Entities.Abstractions;
-
-public interface IEntity
-{
-    public int Id { get; set; }
-}
