--- Data/Infrastructure/ApplicationDbContext.cs
@@ -1,11 +1,0 @@
-using Microsoft.EntityFrameworkCore;
-
-namespace Data.Infrastructure;
-
-public class ApplicationDbContext(IConfigureDbContext configureDbContext) : DbContext
-{
-    // NOTE: fields initialized with primary contsructore are NOT readonly!
-
-    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
-        => configureDbContext.OnDbContextConfiguring(optionsBuilder);
-}
