+++ TestingConsoleApp/ServiceCollectionExtensions.cs
@@ -6,5 +6,5 @@ namespace TestingConsoleApp;
 public static class ServiceCollectionExtensions
 {
     public static IServiceCollection AddServices(this IServiceCollection services)
-        => services.AddBusiness<TestConfigureDbContext>();
+        => services.AddBusiness<TestDbContextConfigurator>();
 }
