+++ WF.Data.Relational/Configurators/BaseEntityConfigurator.cs
@@ -1,16 +1,20 @@
-using Data.Entities.Abstractions;
+using Humanizer;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
+using WF.Data.Relational.Entities;
 
-namespace Data.Configurators.Abstractions;
+namespace WF.Data.Relational.Configurators;
 
-public class BaseConfigurator<TEntity> : IEntityTypeConfiguration<TEntity>
+public abstract class BaseEntityConfigurator<TEntity> : IEntityTypeConfiguration<TEntity>
     where TEntity : class, IEntity
 {
+    protected virtual string TableName => typeof(TEntity).Name.Pluralize();
+    protected virtual string? SchemaName => null;
+
     public void Configure(EntityTypeBuilder<TEntity> builder)
     {
         builder
-            .ToTable(nameof(TEntity))
+            .ToTable(TableName, SchemaName)
             .HasKey(e => e.Id);
 
         ConfigureEntity(builder);
