+++ WF.Data.Relational/Context/ApplicationDbContext.cs
@@ -1,0 +1,9 @@
+using Microsoft.EntityFrameworkCore;
+
+namespace WF.Data.Relational.Context;
+
+public class ApplicationDbContext(IDbContextConfigurator dbContextConfigurator) : DbContext
+{
+    protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        => dbContextConfigurator.OnDbContextConfiguring(optionsBuilder);
+}
