+++ WF.Data.Relational/Entities/BaseEntity.cs
@@ -1,0 +1,6 @@
+namespace WF.Data.Relational.Entities;
+
+public record BaseEntity : IEntity
+{
+    public Guid Id { get; set; }
+}
