+++ WF.Data.Relational/Entities/IEntity.cs
@@ -1,0 +1,6 @@
+namespace WF.Data.Relational.Entities;
+
+public interface IEntity
+{
+    public Guid Id { get; set; }
+}
