+++ WF.Domain/Dtos/BaseDto.cs
@@ -1,0 +1,6 @@
+namespace WF.Domain.Dtos;
+
+public record BaseDto : IDto
+{
+    public Guid Id { get; set; }
+}
