+++ WF.Domain/Dtos/IDto.cs
@@ -1,0 +1,6 @@
+namespace WF.Domain.Dtos;
+
+public interface IDto
+{
+    public Guid Id { get; set; }
+}
